Refer potential users to Trammel
The sentence "Based on Trammel and clojure-contracts" and the fact that
this is a Clojure core library made me think that this is _the_ Clojure
library to use for contracts. However, after some fooling around,
swearing, noticing the need for a `defcontract` macro and starting to
implement my own, I noticed that Trammel is actually better usable.

Add a little note in order to prevent others from making the same
mistake.